登录功能测试成功,前端成功接受数据并传输

+++ src/main/java/com/logisticsystembackend/entity/Car.java
@@ -23,7 +23,7 @@ public class Car implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "car_id", type = IdType.AUTO)
-    private Integer carId;
+    private Long carId;
 
     private String carplate;
