+++ src/main/java/com/logisticsystembackend/entity/Order.java
@@ -25,9 +25,9 @@ public class Order implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "order_id", type = IdType.AUTO)
-    private Integer orderId;
+    private Long orderId;
 
-    private Integer customerId;
+    private Long customerId;
 
     private Integer weight;
